Fix PR comments: add Notification text field, move text logic to mapper, optimize countUnreadNotifications
