+++ service-api/src/main/java/greencity/dto/notification/NewReplyNotificationDto.java
@@ -3,6 +3,7 @@ package greencity.dto.notification;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import greencity.dto.econewscomment.EcoNewsCommentDto;
 import greencity.dto.user.UserVO;
+import greencity.util.NotificationTextFormatter;
 import lombok.Builder;
 
 @Builder
@@ -17,4 +18,15 @@ public record NewReplyNotificationDto(
     public long objectId() {
         return comment.getId();
     }
+
+    @Override
+    public String text() {
+        return NotificationTextFormatter.formatNewReplyText(
+                comment.getAuthor().getName(),
+                "news",
+                comment.getEcoNewsTitle(),
+                comment.getModifiedDate()
+        );
+    }
+
 }
