+++ service-api/src/main/java/greencity/service/NotificationService.java
@@ -1,7 +1,7 @@
 package greencity.service;
 
 import greencity.dto.notification.NewReplyNotificationDto;
-import greencity.dto.notification.NotificationInternalDto;
+//import greencity.dto.notification.NotificationInternalDto;
 import greencity.dto.notification.NotificationDto;
 
 import java.util.List;
@@ -54,7 +54,6 @@ import java.util.List;
         void deleteNotification(Long notificationId);
 
         int countUnreadNotifications(Long userId);
-
 
     }
