+++ service-api/src/main/java/greencity/util/NotificationTextFormatter.java
