+++ service/src/main/java/greencity/mapping/NotificationMapper.java
@@ -4,8 +4,9 @@ package greencity.mapping;
 import greencity.dto.notification.NewReplyNotificationDto;
 import greencity.dto.notification.NotificationDto;
 import greencity.entity.Notification;
-import greencity.enums.NotificationType;
 
+
+import greencity.util.NotificationTextFormatter;
 
 public class NotificationMapper {
     public static Notification mapToEntity(NotificationDto dto) {
@@ -14,10 +15,20 @@ public class NotificationMapper {
         }
 
         if (dto instanceof NewReplyNotificationDto replyDto) {
-            return new NewReplyNotificationDtoMapper().convert(replyDto);
+            Notification notification = new NewReplyNotificationDtoMapper().convert(replyDto);
+
+            // додаємо згенерований текст
+            String text = NotificationTextFormatter.formatNewReplyText(
+                    replyDto.comment().getAuthor().getName(),
+                    "news",
+                    replyDto.comment().getEcoNewsTitle(),
+                    replyDto.comment().getModifiedDate()
+            );
+
+            notification.setText(text);
+            return notification;
         }
 
         throw new IllegalArgumentException("Unsupported NotificationDto type: " + dto.getClass());
     }
 }
-
